+++ minecraft_launcher/__init__.py
@@ -1,0 +1,3 @@
+"""
+Minecraft Libreria para ejecutar Minecraft
+"""
