--- pyproject.toml
@@ -1,41 +1,0 @@
-[tool.poetry]
-name = "xlauncher"
-version = "0.1.0"
-description = "Fully customizable Minecraft launcher"
-authors = [
-    "Keima Senpai <you@example.com>"
-]
-readme = "README.md"
-
-[tool.poetry.dependencies]
-python = ">=3.8,<3.13"
-flet = "^0.21.2"
-flet-route = "^0.3.2"
-pyinstaller = "^6.5.0"
-minecraft-launcher-lib = "^6.4"
-pillow = "^10.3.0"
-
-[tool.poetry.group.dev.dependencies]
-black = { extras = ["d"], version = "^22.3.0" }
-isort = "^5.10.1"
-flake8 = "^5.0.4"
-pre-commit = "^2.18.1"
-pylint = "^2.13.7"
-
-[tool.black]
-target-version = ['py310']
-exclude-extend = '''
-/venv
-/.venv
-/__pycache__
-/.mypy_cache
-/.pytest_cache
-/dist
-/static
-/assets
-/docs
-'''
-
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
