--- requirements-dev.txt
@@ -1,92 +1,0 @@
-aiohttp==3.9.3 ; python_version >= "3.8" and python_version < "3.13"
-aiosignal==1.3.1 ; python_version >= "3.8" and python_version < "3.13"
-altgraph==0.17.4 ; python_version >= "3.8" and python_version < "3.13"
-annotated-types==0.6.0 ; python_version >= "3.8" and python_version < "3.13"
-anyio==4.3.0 ; python_version >= "3.8" and python_version < "3.13"
-arrow==1.3.0 ; python_version >= "3.8" and python_version < "3.13"
-astroid==2.15.8 ; python_version >= "3.8" and python_version < "3.13"
-async-timeout==4.0.3 ; python_version >= "3.8" and python_version < "3.11"
-attrs==23.2.0 ; python_version >= "3.8" and python_version < "3.13"
-binaryornot==0.4.4 ; python_version >= "3.8" and python_version < "3.13"
-black[d]==22.12.0 ; python_version >= "3.8" and python_version < "3.13"
-certifi==2024.2.2 ; python_version >= "3.8" and python_version < "3.13"
-cfgv==3.4.0 ; python_version >= "3.8" and python_version < "3.13"
-chardet==5.2.0 ; python_version >= "3.8" and python_version < "3.13"
-charset-normalizer==3.3.2 ; python_version >= "3.8" and python_version < "3.13"
-click==8.1.3 ; python_version >= "3.8" and python_version < "3.13"
-colorama==0.4.6 ; python_version >= "3.8" and python_version < "3.13" and (sys_platform == "win32" or platform_system == "Windows")
-cookiecutter==2.6.0 ; python_version >= "3.8" and python_version < "3.13"
-dill==0.3.8 ; python_version >= "3.8" and python_version < "3.13"
-distlib==0.3.8 ; python_version >= "3.8" and python_version < "3.13"
-exceptiongroup==1.2.0 ; python_version >= "3.8" and python_version < "3.11"
-fastapi==0.110.1 ; python_version >= "3.8" and python_version < "3.13"
-filelock==3.13.3 ; python_version >= "3.8" and python_version < "3.13"
-flake8==5.0.4 ; python_version >= "3.8" and python_version < "3.13"
-flet-core==0.21.2 ; python_version >= "3.8" and python_version < "3.13"
-flet-route==0.3.2 ; python_version >= "3.8" and python_version < "3.13"
-flet-runtime==0.21.2 ; python_version >= "3.8" and python_version < "3.13"
-flet==0.21.2 ; python_version >= "3.8" and python_version < "3.13"
-frozenlist==1.4.1 ; python_version >= "3.8" and python_version < "3.13"
-h11==0.14.0 ; python_version >= "3.8" and python_version < "3.13"
-httpcore==1.0.5 ; python_version >= "3.8" and python_version < "3.13"
-httptools==0.6.1 ; python_version >= "3.8" and python_version < "3.13"
-httpx==0.27.0 ; python_version >= "3.8" and python_version < "3.13"
-identify==2.5.35 ; python_version >= "3.8" and python_version < "3.13"
-idna==3.6 ; python_version >= "3.8" and python_version < "3.13"
-importlib-metadata==7.1.0 ; python_version >= "3.8" and python_version < "3.10"
-isort==5.13.2 ; python_version >= "3.8" and python_version < "3.13"
-jinja2==3.1.3 ; python_version >= "3.8" and python_version < "3.13"
-lazy-object-proxy==1.10.0 ; python_version >= "3.8" and python_version < "3.13"
-macholib==1.16.3 ; python_version >= "3.8" and python_version < "3.13" and sys_platform == "darwin"
-markdown-it-py==3.0.0 ; python_version >= "3.8" and python_version < "3.13"
-markupsafe==2.1.5 ; python_version >= "3.8" and python_version < "3.13"
-mccabe==0.7.0 ; python_version >= "3.8" and python_version < "3.13"
-mdurl==0.1.2 ; python_version >= "3.8" and python_version < "3.13"
-minecraft-launcher-lib==6.4 ; python_version >= "3.8" and python_version < "3.13"
-multidict==6.0.5 ; python_version >= "3.8" and python_version < "3.13"
-mypy-extensions==1.0.0 ; python_version >= "3.8" and python_version < "3.13"
-nodeenv==1.8.0 ; python_version >= "3.8" and python_version < "3.13"
-oauthlib==3.2.2 ; python_version >= "3.8" and python_version < "3.13"
-packaging==23.2 ; python_version >= "3.8" and python_version < "3.13"
-pathspec==0.12.1 ; python_version >= "3.8" and python_version < "3.13"
-pefile==2023.2.7 ; python_version >= "3.8" and python_version < "3.13" and sys_platform == "win32"
-pillow==10.3.0 ; python_version >= "3.8" and python_version < "3.13"
-platformdirs==4.2.0 ; python_version >= "3.8" and python_version < "3.13"
-pre-commit==2.21.0 ; python_version >= "3.8" and python_version < "3.13"
-pycodestyle==2.9.1 ; python_version >= "3.8" and python_version < "3.13"
-pydantic-core==2.16.3 ; python_version >= "3.8" and python_version < "3.13"
-pydantic==2.6.4 ; python_version >= "3.8" and python_version < "3.13"
-pyflakes==2.5.0 ; python_version >= "3.8" and python_version < "3.13"
-pygments==2.17.2 ; python_version >= "3.8" and python_version < "3.13"
-pyinstaller-hooks-contrib==2024.3 ; python_version >= "3.8" and python_version < "3.13"
-pyinstaller==6.5.0 ; python_version >= "3.8" and python_version < "3.13"
-pylint==2.17.7 ; python_version >= "3.8" and python_version < "3.13"
-pypng==0.20220715.0 ; python_version >= "3.8" and python_version < "3.13"
-python-dateutil==2.9.0.post0 ; python_version >= "3.8" and python_version < "3.13"
-python-dotenv==1.0.1 ; python_version >= "3.8" and python_version < "3.13"
-python-slugify==8.0.4 ; python_version >= "3.8" and python_version < "3.13"
-pywin32-ctypes==0.2.2 ; python_version >= "3.8" and python_version < "3.13" and sys_platform == "win32"
-pyyaml==6.0.1 ; python_version >= "3.8" and python_version < "3.13"
-qrcode==7.4.2 ; python_version >= "3.8" and python_version < "3.13"
-repath==0.9.0 ; python_version >= "3.8" and python_version < "3.13"
-requests==2.31.0 ; python_version >= "3.8" and python_version < "3.13"
-rich==13.7.1 ; python_version >= "3.8" and python_version < "3.13"
-setuptools==69.2.0 ; python_version >= "3.8" and python_version < "3.13"
-six==1.16.0 ; python_version >= "3.8" and python_version < "3.13"
-sniffio==1.3.1 ; python_version >= "3.8" and python_version < "3.13"
-starlette==0.37.2 ; python_version >= "3.8" and python_version < "3.13"
-text-unidecode==1.3 ; python_version >= "3.8" and python_version < "3.13"
-tomli==2.0.1 ; python_version >= "3.8" and python_full_version < "3.11.0a7"
-tomlkit==0.12.4 ; python_version >= "3.8" and python_version < "3.13"
-types-python-dateutil==2.9.0.20240316 ; python_version >= "3.8" and python_version < "3.13"
-typing-extensions==4.10.0 ; python_version >= "3.8" and python_version < "3.13"
-urllib3==2.2.1 ; python_version >= "3.8" and python_version < "3.13"
-uvicorn[standard]==0.29.0 ; python_version >= "3.8" and python_version < "3.13"
-uvloop==0.19.0 ; (sys_platform != "win32" and sys_platform != "cygwin") and platform_python_implementation != "PyPy" and python_version >= "3.8" and python_version < "3.13"
-virtualenv==20.25.1 ; python_version >= "3.8" and python_version < "3.13"
-watchdog==4.0.0 ; python_version >= "3.8" and python_version < "3.13"
-watchfiles==0.21.0 ; python_version >= "3.8" and python_version < "3.13"
-websockets==12.0 ; python_version >= "3.8" and python_version < "3.13"
-wrapt==1.16.0 ; python_version >= "3.8" and python_version < "3.13"
-yarl==1.9.4 ; python_version >= "3.8" and python_version < "3.13"
-zipp==3.18.1 ; python_version >= "3.8" and python_version < "3.10"
