--- tox.ini
@@ -1,12 +1,0 @@
-[flake8]
-max-line-length = 119
-# extend-ignore:
-# E203 is to live with black and do not conflict
-# E266 too many leading '#' for block comment
-# F401 'django.http.request' imported but unused
-# F403 from .celery_callables import *
-# F405 vars may be undefined, or defined from star imports *
-# E501 line too long ( > 119 characters)
-extend-ignore = E203, E266, F401, F403, F405, E501
-count = True
-statistics = True
