--- xlauncher/__init__.py
@@ -1,3 +1,0 @@
-"""
-Minecraft Libreria para ejecutar Minecraft
-"""
